Added Cast button
Made changes to make to cast button appear when a cast device is
available bu adding an item to the menue/browse xml file. Once the
button is clicked the user is able to launch the default cast screen
on a casting device. This was done by connecting the button to a
CastContext.